Removed Console Improved features - Author doesn't want to merge
(I misunderstood them)
